Added files via upload
Create files, begin writing user interface, and copy and modify non-UI code from  particle_path_sim.py (a script I had written to simulate the path of a proton in a magnetic field, but did not create a UI for).  (At the time of this commit, particle_path_sim.py is not on GitHub).